Finalisation des tests

+++ routes/api.php
@@ -20,7 +20,7 @@ Route::group(['middleware' => 'throttle:5,1'], function () {
 Route::group(['middleware' => ['throttle:60,1', 'auth:sanctum']], function () {
     Route::post('/films', 'App\Http\Controllers\FilmController@store')->middleware('role:admin');
     Route::put('/films/{id}', 'App\Http\Controllers\FilmController@update')->middleware('role:admin');
-    Route::delete('/films/{id}', 'App\Http\Controllers\FilmController@destroy')->middleware('check_is_film_deletable');
+    Route::delete('/films/{id}', 'App\Http\Controllers\FilmController@destroy')->middleware('role:admin')->middleware('check_is_film_deletable');
 
     Route::post('/films/{film_id}/critics', 'App\Http\Controllers\CriticController@store')->middleware('critic_limit');
